refactor(opensearch): split opensearch_service.py into a 13-module package
Pure mechanical extraction for #284 A3.5 — no behaviour change. All 62
top-level functions move verbatim (verified AST-identical against HEAD) from
the 3670-line flat module into app/services/opensearch_service/:

  client               the OpenSearch singleton + low-level primitives
  aliases / indices    speakers alias plane, index bootstrap
  repair               corruption detection, repair, rebuild
  transcripts          transcript doc indexing + full-text search
  speaker_write/_read/_maintenance/_metadata/_collections
  profiles / clusters / matching    the kNN read paths

__init__ re-exports every name the flat module exported, so all ~45 importers
keep working unchanged. opensearch_client is resolved through a module-level
__getattr__ rather than a frozen re-export: get_opensearch_client() rebinds it
on lazy re-init, and submodules read it as client.opensearch_client so that
late binding stays visible everywhere.

All 6 cosinesimil->raw-cosine conversion sites that moved
(2.0 * hit["_score"] - 1.0) are unchanged; the repo-wide count is still 11.

Two tests in test_speaker_tenant_gating.py patched module globals on the flat
module and now patch the owning submodule instead.
